Initial work --- tested in console
