+++ NET/API/Treehopper.Desktop/WinUsb/WinUsbConnectionService.cs
@@ -9,9 +9,11 @@ using System.Security.Permissions;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Windows.Forms;
 using Microsoft.Win32.SafeHandles;
 using Treehopper.Desktop.WinUsb;
+using WinApi.Windows;
+using WinApi.Windows.Controls;
+using WinApi.User32;
 
 /// <summary>
 /// Win32-based Treehopper library for Windows desktop support
@@ -25,13 +27,13 @@ namespace Treehopper.Desktop.WinUsb
         public WinUsbConnectionService()
         {
             add($"vid_{TreehopperUsb.Settings.Vid:x}&pid_{TreehopperUsb.Settings.Pid:x}");
-
+            var wf = WindowFactory.Create();
             // We can only hear WM_DEVICECHANGE messages if we're an STA thread that's properly pumping windows messages.
             // There's no easy way to tell if the calling thread is pumping messages, so just check the apartment state, and assume people
             // aren't creating STA threads without a message pump.
             if (Thread.CurrentThread.GetApartmentState() == ApartmentState.STA)
             {
-                mNotifyWindow = new UsbNotifyWindow(this);
+                
             }
             else
             {
@@ -39,17 +41,21 @@ namespace Treehopper.Desktop.WinUsb
                 // Application.Run() to start pumping windows messages.
                 var staThread = new Thread(() =>
                 {
-                    mNotifyWindow = new UsbNotifyWindow(this);
-                    Application.Run();
-                });
 
-                staThread.SetApartmentState(ApartmentState.STA);
-                staThread.Name = "DevNotifyNativeWindow STA Thread";
+                    var mNotifyWindow = wf.CreateWindowEx(() => new UsbNotifyWindow(this), null);
+                    mNotifyWindow.Show();
+                    new EventLoop().Run(mNotifyWindow);
+                })
+                {
+
+                    //staThread.SetApartmentState(ApartmentState.STA);
+                    Name = "DevNotifyNativeWindow STA Thread"
+                };
                 staThread.Start();
             }
         }
 
-        private void add(string matchDevicePath)
+        internal void add(string matchDevicePath)
         {
             //Create structs to hold interface information
             SP_DEVINFO_DATA devInfoData = new SP_DEVINFO_DATA();
@@ -103,144 +109,6 @@ namespace Treehopper.Desktop.WinUsb
         public override void Dispose()
         {
             throw new NotImplementedException();
-        }
-
-        internal sealed class UsbNotifyWindow : NativeWindow
-        {
-            private const string WINDOW_CAPTION = "{18662f14-0871-455c-bf99-eff135425e3a}";
-            private const int WM_DEVICECHANGE = 0x219;
-
-            private const int DBT_DEVICEARRIVAL = 0x8000;
-            private const int DBT_DEVICEREMOVECOMPLETE = 0x8004;
-
-            private const int DBT_DEVTYP_DEVICEINTERFACE = 0x00000005;
-            private readonly DevBroadcastHdr mDevInterface = new DevBroadcastHdr();
-
-            private readonly WinUsbConnectionService service;
-            private DevNotifySafeHandle mDevInterfaceHandle;
-
-            internal UsbNotifyWindow(WinUsbConnectionService winUsbConnectionService)
-            {
-                var cp = new CreateParams();
-                cp.Caption = WINDOW_CAPTION;
-                cp.X = -100;
-                cp.Y = -100;
-                cp.Width = 50;
-                cp.Height = 50;
-                CreateHandle(cp);
-                service = winUsbConnectionService;
-            }
-
-            protected override void OnHandleChange()
-            {
-                if (mDevInterfaceHandle != null)
-                {
-                    mDevInterfaceHandle.Dispose();
-                    mDevInterfaceHandle = null;
-                }
-                if (Handle != IntPtr.Zero)
-                {
-                    mDevInterfaceHandle = NativeMethods.RegisterDeviceNotification(Handle, mDevInterface, 0);
-                    if (mDevInterfaceHandle == null || mDevInterfaceHandle.IsInvalid)
-                    {
-                        Debug.WriteLine($"Device notification register failed with error {Marshal.GetLastWin32Error()}");
-                    }
-                }
-                base.OnHandleChange();
-            }
-
-            [PermissionSet(SecurityAction.Demand, Name = "FullTrust")]
-            protected override void WndProc(ref Message m)
-            {
-                if (m.Msg == WM_DEVICECHANGE)
-                {
-                    var theEvent = m.WParam.ToInt64();
-                    if (theEvent == DBT_DEVICEARRIVAL || theEvent == DBT_DEVICEREMOVECOMPLETE)
-                    {
-                        var msg = new DevBroadcastDeviceInterface();
-                        Marshal.PtrToStructure(m.LParam, msg);
-                        if (msg.DeviceType != 0x05)
-                            return;
-                        if (theEvent == DBT_DEVICEARRIVAL)
-                        {
-                            Task.Run(() => { service.add(msg.DevicePath); });
-                        }
-                        else
-                        {
-                            var board = service.Boards
-                                .Where(x => x.Connection.DevicePath.Substring(3).ToLower() ==
-                                            msg.DevicePath.Substring(3).ToLower())
-                                .FirstOrDefault();
-                            if (board != null)
-                            {
-                                Debug.WriteLine("Removing: " + board);
-                                board.Connection.Dispose(); // kill the connection first
-                                board.Dispose();
-                                service.Boards.Remove(board);
-                            }
-                        }
-                    }
-                }
-
-                base.WndProc(ref m);
-            }
-
-            internal static class NativeMethods
-            {
-                [DllImport("user32.dll", SetLastError = true, EntryPoint = "RegisterDeviceNotificationA",
-                    CharSet = CharSet.Ansi)]
-                internal static extern DevNotifySafeHandle RegisterDeviceNotification(
-                    IntPtr hRecipient,
-                    [MarshalAs(UnmanagedType.AsAny)] [In] object notificationFilter,
-                    int flags);
-
-                [DllImport("user32.dll", SetLastError = true)]
-                internal static extern bool UnregisterDeviceNotification(IntPtr handle);
-            }
-
-            [StructLayout(LayoutKind.Sequential)]
-            internal class DevBroadcastHdr
-            {
-                public int Size;
-                public int DeviceType = DBT_DEVTYP_DEVICEINTERFACE;
-                public int Reserved;
-                public Guid UsbDevice = new Guid("A5DCBF10-6530-11D2-901F-00C04FB951ED");
-                private char mNameHolder;
-
-                internal DevBroadcastHdr()
-                {
-                    Size = Marshal.SizeOf(this);
-                }
-            }
-
-
-            [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
-            public class DevBroadcastDeviceInterface
-            {
-                public int Size;
-                public int DeviceType;
-                public int Reserved;
-                public Guid ClassGuid;
-                [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 255)] public string DevicePath;
-            }
-
-            internal class DevNotifySafeHandle : SafeHandleZeroOrMinusOneIsInvalid
-            {
-                public DevNotifySafeHandle() : base(true)
-                {
-                }
-
-                protected override bool ReleaseHandle()
-                {
-                    if (handle != IntPtr.Zero)
-                    {
-                        var bSuccess = NativeMethods.UnregisterDeviceNotification(handle);
-                        handle = IntPtr.Zero;
-                    }
-
-                    return true;
-                }
-            }
         }
     }
 }
