+++ NET/Demos/Console/BlinkCore/BlinkCore.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\API\Treehopper.Desktop\Treehopper.Desktop.csproj" />
+    <ProjectReference Include="..\..\..\API\Treehopper\Treehopper.csproj" />
+  </ItemGroup>
+
+</Project>
